+++ src/main/java/net/swordie/ms/client/jobs/nova/AngelicBuster.java
@@ -501,7 +501,7 @@ public class AngelicBuster extends Job {
                 affinityHeartIIcounter = 0;
                 affinityHeartIIIcounter = 0;
             } else {
-                int totaldmg = Arrays.stream(mai.damages).sum();
+                long totaldmg = Arrays.stream(mai.damages).sum();
                 if (totaldmg > mob.getHp()) {
                     affinityHeartIIcounter++;
                 }
